Refactors Styles and adds Home and Header components
