+++ src/components/Header/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Header = () => (
+  <h1>This is the Pokemon header</h1>
+)
+
+export default Header;
