+++ src/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import 'bootstrap/dist/css/bootstrap.css';
+import Home from './components/Home';
+import registerServiceWorker from './registerServiceWorker';
+
+ReactDOM.render(<Home />, document.getElementById('root'));
+registerServiceWorker();
