+++ src/styles/components/PokemonType/PokemonType.css
